pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "esloco"
version = "1.0.5"
authors = [{name = "aweich"}]
description = "Estimating Long-Read Coverage for Whole-Genome Sequencing"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "biopython",
    "numpy",
    "pandas",
    "joblib",
    "tqdm_joblib",
    "psutil",
    "matplotlib",
    "seaborn",
    "scipy",
    "plotly",
    "pyfiglet",
    "kaleido<1.0.0", #for svg image export
]

[project.scripts]
esloco = "esloco.esloco:main"
plot_esloco = "esloco.plot_esloco:main"

[project.urls]
Repository = "https://github.com/aweich/esloco.git"
